back/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
astunparse==1.6.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
Django==5.1.1
fastapi==0.115.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.66.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
image==1.5.33
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.5.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.4.0
msgpack==1.0.7
namex==0.0.8
numpy==1.24.3
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pillow==10.4.0
protobuf==4.25.4
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pynvim==0.4.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
rich==13.8.0
six==1.16.0
sniffio==1.3.1
spotipy==2.24.0
sqlparse==0.5.1
starlette==0.38.6
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-keras-lite==0.1.0
termcolor==2.4.0
tk==0.1.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
websockets==13.0.1
Werkzeug==3.0.4
wrapt==1.16.0
